PANA-5681: Move isValidTapTarget to better location
