+++ dd-sdk-android-internal/api/apiSurface
@@ -3,7 +3,6 @@ data class com.datadog.android.heatmaps.CrossPlatformHeatmapActionData
 interface com.datadog.android.heatmaps.HeatmapIdentifierRegistryProvider
   val heatmapIdentifierRegistry: com.datadog.android.internal.heatmaps.HeatmapIdentifierRegistry
 fun heatmapViewKey(android.view.View): Long
-fun android.view.View.isValidTapTarget(): Boolean
 interface com.datadog.android.internal.attributes.LocalAttribute
   enum Key
     constructor(String)
@@ -285,6 +284,7 @@ fun StringBuilder.appendIfNotEmpty(String)
 fun StringBuilder.appendIfNotEmpty(Char)
 fun String.toBase64(): String
 fun String.fromBase64(): String?
+fun android.view.View.isValidTapTarget(): Boolean
 fun Thread.safeGetThreadId(): Long
 fun Thread.State.asString(): String
 fun Array<StackTraceElement>.loggableStackTrace(): String
