+++ dd-sdk-android-internal/src/test/java/com/datadog/android/internal/utils/TapTargetUtilsTest.kt
@@ -4,7 +4,7 @@
  * Copyright 2016-Present Datadog, Inc.
  */
 
-package com.datadog.android.heatmaps
+package com.datadog.android.internal.utils
 
 import android.view.View
 import com.datadog.android.internal.forge.Configurator
