+++ features/dd-sdk-android-session-replay/src/main/kotlin/com/datadog/android/sessionreplay/internal/recorder/HeatmapIdentifierResolver.kt
@@ -13,9 +13,9 @@ import androidx.annotation.UiThread
 import androidx.annotation.VisibleForTesting
 import com.datadog.android.api.InternalLogger
 import com.datadog.android.heatmaps.heatmapViewKey
-import com.datadog.android.heatmaps.isValidTapTarget
 import com.datadog.android.internal.heatmaps.HeatmapIdentifier
 import com.datadog.android.internal.heatmaps.HeatmapIdentifierRegistry
+import com.datadog.android.internal.utils.isValidTapTarget
 
 internal class HeatmapIdentifierResolver(
     private val appPackageName: String,
